Refactor title component and import statements
