+++ src/app/main/booking/request/components/dry-container-input.tsx
@@ -18,6 +18,7 @@ import { Disclosure } from "@headlessui/react";
 import { Add, ArrowDropDown, DeleteOutline } from "@mui/icons-material";
 
 import DangerousCargoInput from "./dangerous-cargo-input";
+import { DetailTitle } from "@/app/components/title-components";
 
 const DryContainerInputContainer = ({
   list,
@@ -31,10 +32,7 @@ const DryContainerInputContainer = ({
       {({ open }) => (
         <>
           <Disclosure.Button className={`flex w-full items-center gap-2`}>
-            <div className="w-1 h-4 bg-primary"></div>
-            <MdTypography variant="body" size="large" prominent>
-              Dry Container
-            </MdTypography>
+            <DetailTitle title="Dry Container" />
             <div className="flex-1 border-b border-b-outlineVariant"></div>
             <ArrowDropDown
               className={`transform transition-transform ${
