+++ src/app/main/booking/request/components/flatrack-container-input.tsx
@@ -17,6 +17,7 @@ import { Add, ArrowDropDown, DeleteOutline } from "@mui/icons-material";
 import { useSetRecoilState } from "recoil";
 import DangerousCargoInput from "./dangerous-cargo-input";
 import AwkwardContainerInput from "./awkward-container-input";
+import { DetailTitle } from "@/app/components/title-components";
 
 const FlatRackContainerInput = ({
   list,
@@ -31,10 +32,7 @@ const FlatRackContainerInput = ({
         return (
           <>
             <Disclosure.Button className={`flex w-full items-center gap-2`}>
-              <div className="w-1 h-4 bg-primary"></div>
-              <MdTypography variant="body" size="large" prominent>
-                Flat Rack Container
-              </MdTypography>
+              <DetailTitle title="Flat Rack Container" />
               <div className="flex-1 border-b border-b-outlineVariant"></div>
               <ArrowDropDown
                 className={`transform transition-transform ${
