+++ src/app/main/booking/request/components/opentop-container-input.tsx
@@ -16,6 +16,7 @@ import { Add, ArrowDropDown, DeleteOutline } from "@mui/icons-material";
 import { useSetRecoilState } from "recoil";
 import DangerousCargoInput from "./dangerous-cargo-input";
 import AwkwardContainerInput from "./awkward-container-input";
+import { DetailTitle } from "@/app/components/title-components";
 
 const OpenTopContainerInput = ({
   list,
@@ -30,10 +31,7 @@ const OpenTopContainerInput = ({
         return (
           <>
             <Disclosure.Button className={`flex w-full items-center gap-2`}>
-              <div className="w-1 h-4 bg-primary"></div>
-              <MdTypography variant="body" size="large" prominent>
-                Open Top Container
-              </MdTypography>
+              <DetailTitle title="Open Top Container" />
               <div className="flex-1 border-b border-b-outlineVariant"></div>
               <ArrowDropDown
                 className={`transform transition-transform ${
