+++ src/app/main/booking/request/components/reefer-container-input.tsx
@@ -15,6 +15,7 @@ import { Add, ArrowDropDown, DeleteOutline } from "@mui/icons-material";
 import { useSetRecoilState } from "recoil";
 import DangerousCargoInput from "./dangerous-cargo-input";
 import NAOutlinedListBox from "@/app/components/na-outline-listbox";
+import { DetailTitle } from "@/app/components/title-components";
 
 const ReeferContainerInput = ({
   list,
@@ -29,10 +30,7 @@ const ReeferContainerInput = ({
         return (
           <>
             <Disclosure.Button className={`flex w-full items-center gap-2`}>
-              <div className="w-1 h-4 bg-primary"></div>
-              <MdTypography variant="body" size="large" prominent>
-                Reefer Container
-              </MdTypography>
+              <DetailTitle title="Reefer Container" />
               <div className="flex-1 border-b border-b-outlineVariant"></div>
               <ArrowDropDown
                 className={`transform transition-transform ${
